Implement new requested features
This adds the ability to add points to map with editable lat/lng and current weather.

Note: this is an usually large commit, but contains many interrelated changes that would be difficult to chunk out into smaller (still meaningful) commits. One of the big changes was the introduction of Redux to manage global state. This helped facilitate data binding between the map, markers, popups, and lng/lat labels.